+++ tasks/seq/vinicuk_t_complex_sparse_matrix_mult_csc/perf_tests/main.cpp
@@ -14,7 +14,7 @@ TEST(vinicuk_t_complex_sparse_matrix_mult_csc_seq, test_pipeline_run_dft384x384)
   CSCComplexMatrix mtrx_res(1, n);
 
   int k = 1, p = 0;
-  for (int i = 0; i <= n; i++) {
+  for (size_t i = 0; i <= n; i++) {
     mtrx_A.col_ptrs.push_back(p);
     p += k;
     k++;
@@ -22,7 +22,7 @@ TEST(vinicuk_t_complex_sparse_matrix_mult_csc_seq, test_pipeline_run_dft384x384)
 
   k = 1;
   p = 0;
-  for (int i = 0; i < mtrx_A.col_ptrs[n]; i++) {
+  for (size_t i = 0; i < mtrx_A.col_ptrs[n]; i++) {
     mtrx_A.values.push_back(std::complex<double>(2.0, 1.0));
     if (p >= k) {
       p = 0;
@@ -33,7 +33,7 @@ TEST(vinicuk_t_complex_sparse_matrix_mult_csc_seq, test_pipeline_run_dft384x384)
   }
 
   mtrx_B.col_ptrs = {0, n};
-  for (int i = 0; i < n; i++) {
+  for (size_t i = 0; i < n; i++) {
     mtrx_B.values.push_back(std::complex<double>(2.0, 1.0));
     mtrx_B.row_indexes.push_back(i);
   }
@@ -73,7 +73,7 @@ TEST(vinicuk_t_complex_sparse_matrix_mult_csc_seq, test_task_run_dft384x384) {
   CSCComplexMatrix mtrx_res(1, n);
 
   int k = 1, p = 0;
-  for (int i = 0; i <= n; i++) {
+  for (size_t i = 0; i <= n; i++) {
     mtrx_A.col_ptrs.push_back(p);
     p += k;
     k++;
@@ -82,7 +82,7 @@ TEST(vinicuk_t_complex_sparse_matrix_mult_csc_seq, test_task_run_dft384x384) {
 
   k = 1;
   p = 0;
-  for (int i = 0; i < mtrx_A.col_ptrs[n]; i++) {
+  for (size_t i = 0; i < mtrx_A.col_ptrs[n]; i++) {
     mtrx_A.values.push_back(std::complex<double>(2.0, 1.0));
     if (p >= k) {
       p = 0;
@@ -93,7 +93,7 @@ TEST(vinicuk_t_complex_sparse_matrix_mult_csc_seq, test_task_run_dft384x384) {
   }
 
   mtrx_B.col_ptrs = {0, n};
-  for (int i = 0; i < n; i++) {
+  for (size_t i = 0; i < n; i++) {
     mtrx_B.values.push_back(std::complex<double>(2.0, 1.0));
     mtrx_B.row_indexes.push_back(i);
   }
